+++ login-utils-dal/src/main/java/org/loginutils/dal/mappers/RolePermissionMapper.java
@@ -1,0 +1,10 @@
+package org.loginutils.dal.mappers;
+
+import org.apache.ibatis.annotations.Mapper;
+import org.loginutils.dal.model.RolePermissionDo;
+
+@Mapper
+public interface RolePermissionMapper {
+    int insert(RolePermissionDo record);
+    RolePermissionDo selectByPrimaryKey(Long id);
+}