+++ login-utils-dal/src/main/java/org/loginutils/dal/mappers/UserRoleMapper.java
@@ -1,6 +1,7 @@
 package org.loginutils.dal.mappers;
 
 import org.apache.ibatis.annotations.Mapper;
+import org.loginutils.common.dto.UserRoleDetailDto;
 import org.loginutils.dal.model.UserRoleDo;
 
 import java.util.List;
@@ -8,7 +9,7 @@ import java.util.List;
 @Mapper
 public interface UserRoleMapper {
 
-    List<UserRoleDo> findRolesByUserId(Long userId);
+    List<UserRoleDetailDto> findRolesByUserId(Long userId);
 
     void insertUserRole(UserRoleDo userRoleDo);