+++ login-utils-dal/src/main/java/org/loginutils/dal/model/RoleDo.java
@@ -1,0 +1,21 @@
+package org.loginutils.dal.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import java.sql.Timestamp;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class RoleDo {
+    private Long roleId;
+    private Long systemId;
+    private String roleCode;
+    private String roleName;
+    private Integer status;
+    private Timestamp createTime;
+    private Timestamp updateTime;
+}