+++ login-utils-mgr/src/main/java/org/loginutils/mgr/controller/login/LoginResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.loginutils.common.dto.RoleDto;
 
 import java.util.List;
 
@@ -14,5 +15,5 @@ import java.util.List;
 public class LoginResponse {
     private String token;
     private Long userId;
-    private List<String> roles;
+    private List<RoleDto> roles;
 }